+++ src/components/Footer.tsx
@@ -39,33 +39,31 @@ const useStyles = makeStyles({
   },
 });
 
+const menuItems = [
+  {
+    label: "トップページ",
+    href: "/",
+  },
+  {
+    label: "このアプリについて",
+    href: "/about",
+  },
+  {
+    label: "レシピの計算方法について",
+    href: "/calc-procedure",
+  },
+];
+
 export default function Footer(props: Props) {
   const classes = useStyles();
 
-  const menuItems = [
-    {
-      label: "トップページ",
-      href: "/",
-    },
-    {
-      label: "このアプリについて",
-      href: "/about",
-    },
-    {
-      label: "レシピの計算方法について",
-      href: "/calc-procedure",
-    },
-  ];
-
   const menuItemElms = menuItems.map((v) => {
     return (
-      <footer>
-        <li className={classes.menuItem} key={`pagelink_${v.href}`}>
-          <Link to={v.href} className={classes.link}>
-            {v.label}
-          </Link>
-        </li>
-      </footer>
+      <li className={classes.menuItem} key={`pagelink_${v.href}`}>
+        <Link to={v.href} className={classes.link}>
+          {v.label}
+        </Link>
+      </li>
     );
   });
