+++ src/components/NutrientsDetail.tsx
@@ -3,7 +3,7 @@ import Table from "@material-ui/core/Table";
 import TableBody from "@material-ui/core/TableBody";
 import TableCell from "@material-ui/core/TableCell";
 import TableRow from "@material-ui/core/TableRow";
-import { sumUpNutrients, Ingredient } from "../libs/calc";
+import { sumUpNutrients, Ingredient } from "../libs/calculator/calc";
 
 type NutrientsDetailProps = {
   ingredients: Ingredient[];
