+++ src/components/PFCBalanceInput.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Box, InputAdornment, TextField, Typography } from "@material-ui/core";
 import Alert from "@material-ui/lab/Alert";
-import { PFCBalance } from "../libs/calc";
+import { PFCBalance } from "../libs/calculator/calc";
 import { validationErrorMessage } from "../utils/messages";
 import { ValidationFuncs, getValidationMessage } from "../utils/validation";
