+++ src/libs/calculator/foodsData.ts
